put get_flipped() in one file
